extracting more library components
